feat: add handlers and rl events
