+++ src/functions/exit.js
@@ -1,21 +1,5 @@
-import readline from 'readline';
-import { showGoodbye } from './showMeassages.js';
-import { stdin, stdout } from '../utils/common.js';
+// import { showGoodbye } from './showMeassages.js';
 
-const rl = readline.createInterface({
-  input: stdin,
-  output: stdout
-});
+// export const exit = () => {
 
-rl.on('line', (input) => {
-  if (input.trim() === '.exit') {
-    rl.close();
-  } 
-});
-
-export const exit = () => {
-  rl.on('close', () => {
-    showGoodbye();
-    process.exit(0);
-  });
-}
+// }
