+++ src/index.js
@@ -1,8 +1,8 @@
 import { changeWorkDirectory } from "./functions/changeDirectory.js";
-import { exit } from "./functions/exit.js";
-import { showCurrentDirectory, showWelcome } from "./functions/showMeassages.js";
+import { handleEvents } from "./events/handler.js";
+import { START } from "./utils/consts.js";
+import { actionRL } from './events/rl.js';
 
 changeWorkDirectory();
-showWelcome();
-showCurrentDirectory();
-exit();
+handleEvents(START);
+actionRL();
