+++ src/utils/consts.js
@@ -1,0 +1,7 @@
+const START = 'start';
+const INVALID_INPUT = 'Invalid input';
+
+export {
+  START,
+  INVALID_INPUT
+};
